feat(widget-chat): Move components and containers into folders
Move components and containers into folders to reflect pattern in other
projects. Add initial empty Activity Title component.
