+++ packages/widget-chat/src/components/activity-title/activity-title.test.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
+
+import ActivityTitle from './activity-title';
+
+it(`is rendered properly`, () => {
+  const title = TestUtils.renderIntoDocument(
+    <ActivityTitle heading="Chat Widget Title!" />
+  );
+
+  const titleNode = ReactDOM.findDOMNode(title);
+
+  expect(titleNode.textContent).toBe(`Chat Widget Title!`);
+});
