+++ packages/widget-chat/src/components/activity-title/styles.css
@@ -1,0 +1,11 @@
+.activity-title {
+  position: relative;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  padding-left: 80px;
+  padding-right: 25px;
+  background-color: transparent;
+  border-bottom: 1px solid #ccc;
+  z-index: 2;
+}
