+++ packages/widget-chat/src/containers/chat-widget.test.js
