+++ packages/widget-chat/src/containers/styles.css
