+++ packages/widget-chat/src/main.js
@@ -2,7 +2,7 @@ import './styles/styles.css';
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import ChatWidget from './chat-widget/chat-widget';
+import ChatWidget from './containers/chat-widget';
 
 
 ReactDOM.render(
